add tracer_validator class for abstraction
