+++ integration/tests/tracer/test_tracer_debug_trace_complex_transactions.py
@@ -24,56 +24,6 @@ class TestDebugTraceIterativeTransaction:
     accounts: EthAccounts
     tracer_api: TracerClient
 
-    def check_call_tracer_type(
-        self,
-        tx_data,
-        wait_error=False,
-        error_message="",
-    ):
-
-        params = [tx_data["hash"].hex(), {"tracer": "callTracer", "tracerConfig": {"withLog": True}}]
-        response = self.tracer_api.send_rpc_and_wait_response("debug_traceTransaction", params)
-
-        assert response["result"]["from"].lower() == tx_data["from"].lower()
-        assert response["result"]["to"].lower() == tx_data["to"].lower()
-        assert response["result"]["input"].lower() == "0x" + tx_data["input"].hex().lower()
-        assert response["result"]["type"] == "CALL"
-
-        if wait_error:
-            assert "error" in response["result"]
-            assert response["result"]["error"] == error_message
-        else:
-            assert "error" not in response["result"]
-
-        return response
-
-    def check_tracer_struct_log(self, tx_data, wait_error=False, wait_return_value=False, return_value=""):
-        params = [
-            {
-                "to": tx_data["to"],
-                "from": tx_data["from"],
-                "gas": hex(tx_data["gas"]),
-                "gasPrice": hex(tx_data["gasPrice"]),
-                "value": hex(tx_data["value"]),
-                "data": "0x" + tx_data["input"].hex(),
-            },
-            hex(tx_data["blockNumber"]),
-        ]
-
-        response = self.tracer_api.send_rpc_and_wait_response("debug_traceCall", params)
-
-        if wait_error:
-            assert response["result"]["failed"] is True
-        else:
-            assert "error" not in response["result"], "Error in response"
-        if wait_return_value:
-            assert (
-                response["result"]["returnValue"] == return_value
-            ), f'Waited {return_value}, got {response["result"]["returnValue"]}'
-        validate_response_result(response)
-
-        return response
-
     @pytest.mark.skip(reason="NDEV-3595, take NDEV-3611 after the fix")
     def test_trace_iterative_tx_struct_opcode_tracer(self, counter_contract):
         sender_account = self.accounts[0]
@@ -361,7 +311,14 @@ class TestDebugTraceIterativeTransaction:
         self.check_call_tracer_type(tx_data, wait_error=True, error_message="execution reverted")
 
     def test_trace_failed_multiply_scheduled_tx(
-        self, web3_client_sol, neon_user, treasury_pool, revert_contract_caller, event_caller_contract, evm_loader
+        self,
+        web3_client_sol,
+        neon_user,
+        treasury_pool,
+        revert_contract_caller,
+        event_caller_contract,
+        evm_loader,
+        tracer_validator,
     ):
         nonce = web3_client_sol.get_nonce(neon_user.checksum_address)
 
@@ -407,8 +364,8 @@ class TestDebugTraceIterativeTransaction:
 
         receipt = web3_client_sol.wait_for_transaction_receipt(tx0.hash().hex())
         tx_data = self.web3_client.get_transaction_by_hash(receipt["transactionHash"].hex())
-        self.check_tracer_struct_log(tx_data, wait_error=True)
-        self.check_call_tracer_type(tx_data, wait_error=True, error_message="execution reverted")
+        tracer_validator.check_tracer_struct_log(tx_data, wait_error=True)
+        tracer_validator.check_call_tracer_type(tx_data, wait_error=True, error_message="execution reverted")
 
         error_message = "Tracing Skip Scheduled Transaction is not supported"
         params = [tx1.hash().hex(), tracer_params]
