+++ integration/tests/tracer/test_tracer_debug_trace_transactions_callTracer.py
@@ -10,7 +10,6 @@ from deepdiff import DeepDiff
 from solders.pubkey import Pubkey
 
 from integration.tests.basic.helpers.basic import AccountData
-from integration.tests.tracer.tracer_helper import validate_response_result
 from utils.consts import COUNTER_ID, LAMPORT_PER_SOL
 from utils.operator import Operator
 from utils.solana_client import SolanaClient
@@ -106,58 +105,6 @@ class TestDebugTraceTransactionCallTracer:
                 expected_response["calls"][0]["revertReason"] = revert_reason
 
         return expected_response
-
-    def check_call_tracer_type(
-        self,
-        tx_data,
-        wait_error=False,
-        error_message="",
-    ) -> dict:
-
-        params = [tx_data["hash"].hex(), {"tracer": "callTracer", "tracerConfig": {"withLog": True}}]
-        response = self.tracer_api.send_rpc_and_wait_response("debug_traceTransaction", params)
-
-        assert response["result"]["from"].lower() == tx_data["from"].lower()
-        assert response["result"]["to"].lower() == tx_data["to"].lower()
-        assert response["result"]["input"].lower() == "0x" + tx_data["input"].hex().lower()
-        assert response["result"]["type"] == "CALL"
-
-        if wait_error:
-            assert "error" in response["result"]
-            assert response["result"]["error"] == error_message
-        else:
-            assert "error" not in response["result"]
-
-        return response
-
-    def check_tracer_struct_log(
-        self, tx_data, wait_error=False, error_message="", wait_return_value=False, return_value=""
-    ):
-        params = [
-            {
-                "to": tx_data["to"],
-                "from": tx_data["from"],
-                "gas": hex(tx_data["gas"]),
-                "gasPrice": hex(tx_data["gasPrice"]),
-                "value": hex(tx_data["value"]),
-                "data": "0x" + tx_data["input"].hex(),
-            },
-            hex(tx_data["blockNumber"]),
-        ]
-
-        response = self.tracer_api.send_rpc_and_wait_response("debug_traceCall", params)
-
-        if wait_error:
-            assert response["result"]["failed"] is True
-        else:
-            assert "error" not in response["result"], "Error in response"
-        if wait_return_value:
-            assert (
-                response["result"]["returnValue"] == return_value
-            ), f'Waited {return_value}, got {response["result"]["returnValue"]}'
-        validate_response_result(response)
-
-        return response
 
     def assert_response_contains_expected(self, pytestconfig, expected_response, response, sort_calls=False):
         if sort_calls:
@@ -585,6 +532,7 @@ class TestDebugTraceTransactionCallTracer:
         operator: Operator,
         web3_client: NeonChainWeb3Client,
         accounts: EthAccounts,
+        tracer_validator,
     ):
         tx_type = TransactionType(2)
         sender_account = accounts[0]
@@ -604,12 +552,10 @@ class TestDebugTraceTransactionCallTracer:
         receipt = web3_client.send_transaction(sender_account, instruction_tx)
         assert receipt["status"] == 1
 
-        tx_data = self.web3_client.get_transaction_by_hash(receipt["transactionHash"].hex())
-        self.check_tracer_struct_log(tx_data)
-        self.check_call_tracer_type(tx_data)
+        tracer_validator.check_all_tracer_types(receipt)
 
     def test_callTracer_precompiled_solana_contract(
-        self, call_solana_caller, counter_resource_address: bytes, pytestconfig
+        self, call_solana_caller, counter_resource_address: bytes, pytestconfig, tracer_validator
     ):
         sender = self.accounts[0]
         lamports = 0
@@ -629,6 +575,4 @@ class TestDebugTraceTransactionCallTracer:
         )
 
         receipt = self.web3_client.send_transaction(sender, instruction_tx)
-        tx_data = self.web3_client.get_transaction_by_hash(receipt["transactionHash"].hex())
-        self.check_tracer_struct_log(tx_data)
-        self.check_call_tracer_type(tx_data)
+        tracer_validator.check_all_tracer_types(receipt)
